認可

+++ src/app/Http/Controllers/Tweet/DeleteController.php
@@ -4,16 +4,22 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Services\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\HttpFoundation\File\Exception\AccessDeniedException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int)$request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new NotFoundHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()->route('tweet.index')->with('feedback.success', 'Tweet deleted successfully');
